# klickz-backend/cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for gin backend template.",
        "title": "gin-api-template Swagger API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/files/upload/{fileName}": {
            "post": {
                "description": "Send as binary file and write name of the file in the URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Uploading files",
                "parameters": [
                    {
                        "description": "File data",
                        "name": "file",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.File"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.SimpleError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "data.File": {
            "type": "object",
            "properties": {
                "createdTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "main.SimpleError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}